requirements.txt
crewai>=0.75.0
crewai-tools>=0.17.0
gradio>=5.12.0
litellm>=1.50.2
openai>=1.59.6
pydantic>=2.10.5
Pillow>=11.1.0
python-dotenv>=1.0.1
requests>=2.32.0
PyYAML>=6.0.2
